handle lines ending with ‘:’ and preserve surrounding whitespace
make it possible to format a single line such as ‘if True:’, which is
not a complete python program, by adding (and later removing) a ‘pass’
statement at the end.

extract the core logic into a machhiato() function, and call it from
main(), which makes it easier to run tests, which have also been added.